+++ build-lualib.js
@@ -6,11 +6,8 @@ const tstl = require("./src");
 const { loadLuaLibFeatures } = require("./src/LuaLib");
 
 const configFileName = path.resolve(__dirname, "src/lualib/tsconfig.json");
-const { emitResult, diagnostics } = tstl.transpileProject(configFileName);
-emitResult.forEach(({ name, text }) => ts.sys.writeFile(name, text));
-
-const reportDiagnostic = tstl.createDiagnosticReporter(true);
-diagnostics.forEach(reportDiagnostic);
+const { diagnostics } = tstl.transpileProject(configFileName);
+diagnostics.forEach(tstl.createDiagnosticReporter(true));
 
 const bundlePath = path.join(__dirname, "dist/lualib/lualib_bundle.lua");
 if (fs.existsSync(bundlePath)) {