+++ src/transformation/visitors/access.ts
@@ -3,8 +3,9 @@ import * as lua from "../../LuaAST";
 import { transformBuiltinPropertyAccessExpression } from "../builtins";
 import { FunctionVisitor, TransformationContext } from "../context";
 import { AnnotationKind, getTypeAnnotations } from "../utils/annotations";
-import { createExpressionPlusOne } from "../utils/lua-ast";
-import { isArrayType, isNumberType, isStringType, isExpressionWithEvaluationEffect } from "../utils/typescript";
+import { addToNumericExpression } from "../utils/lua-ast";
+import { LuaLibFeature, transformLuaLibFunction } from "../utils/lualib";
+import { isArrayType, isNumberType, isStringType } from "../utils/typescript";
 import { tryGetConstEnumValue } from "./enum";
 import { transformLuaTablePropertyAccessExpression, validateLuaTableElementAccessExpression } from "./lua-table";
 
@@ -16,32 +17,31 @@ export function transformElementAccessArgument(
 
     const type = context.checker.getTypeAtLocation(node.expression);
     const argumentType = context.checker.getTypeAtLocation(node.argumentExpression);
-    if (isNumberType(context, argumentType) && isArrayType(context, type)) {
-        return createExpressionPlusOne(index);
+    if (isArrayType(context, type) && isNumberType(context, argumentType)) {
+        return addToNumericExpression(index, 1);
     }
 
     return index;
 }
 
-export const transformElementAccessExpression: FunctionVisitor<ts.ElementAccessExpression> = (expression, context) => {
-    validateLuaTableElementAccessExpression(context, expression);
+export const transformElementAccessExpression: FunctionVisitor<ts.ElementAccessExpression> = (node, context) => {
+    validateLuaTableElementAccessExpression(context, node);
 
-    const constEnumValue = tryGetConstEnumValue(context, expression);
+    const constEnumValue = tryGetConstEnumValue(context, node);
     if (constEnumValue) {
         return constEnumValue;
     }
 
-    const argumentType = context.checker.getTypeAtLocation(expression.argumentExpression);
-    const type = context.checker.getTypeAtLocation(expression.expression);
-    if (isNumberType(context, argumentType) && isStringType(context, type)) {
-        return transformStringIndex(context, expression);
+    const table = context.transformExpression(node.expression);
+
+    const type = context.checker.getTypeAtLocation(node.expression);
+    const argumentType = context.checker.getTypeAtLocation(node.argumentExpression);
+    if (isStringType(context, type) && isNumberType(context, argumentType)) {
+        const index = context.transformExpression(node.argumentExpression);
+        return transformLuaLibFunction(context, LuaLibFeature.StringAccess, node, table, index);
     }
 
-    return lua.createTableIndexExpression(
-        context.transformExpression(expression.expression),
-        transformElementAccessArgument(context, expression),
-        expression
-    );
+    return lua.createTableIndexExpression(table, transformElementAccessArgument(context, node), node);
 };
 
 export const transformPropertyAccessExpression: FunctionVisitor<ts.PropertyAccessExpression> = (
@@ -91,32 +91,3 @@ export const transformQualifiedName: FunctionVisitor<ts.QualifiedName> = (node, 
 
     return lua.createTableIndexExpression(left, right, node);
 };
-
-function transformStringIndex(context: TransformationContext, expression: ts.ElementAccessExpression): lua.Expression {
-    const string = context.transformExpression(expression.expression);
-    // Translate to string.sub(str, index, index), cache index in case it has side effects.
-    if (isExpressionWithEvaluationEffect(expression.argumentExpression)) {
-        const indexIdentifier = lua.createIdentifier("____index");
-        // string.sub(stringExpression, ____index, ____index)
-        const subCall = lua.createCallExpression(
-            lua.createTableIndexExpression(lua.createIdentifier("string"), lua.createStringLiteral("sub")),
-            [string, lua.cloneIdentifier(indexIdentifier), lua.cloneIdentifier(indexIdentifier)],
-            expression
-        );
-        // function(____index) string.sub(stringExpression, ____index, ____index)
-        const functionExpression = lua.createFunctionExpression(
-            lua.createBlock([lua.createReturnStatement([subCall])]),
-            [lua.cloneIdentifier(indexIdentifier)]
-        );
-        // (function(____index) string.sub(stringExpression, ____index, ____index) end)(index + 1)
-        const indexPlusOne = createExpressionPlusOne(context.transformExpression(expression.argumentExpression));
-        return lua.createCallExpression(functionExpression, [indexPlusOne]);
-    } else {
-        const index = context.transformExpression(expression.argumentExpression);
-        return lua.createCallExpression(
-            lua.createTableIndexExpression(lua.createIdentifier("string"), lua.createStringLiteral("sub")),
-            [string, createExpressionPlusOne(index), createExpressionPlusOne(index)],
-            expression
-        );
-    }
-}