+++ test/transpile/project.spec.ts
@@ -1,18 +1,8 @@
 import * as path from "path";
-import { transpileProject } from "../../src";
-
-const projectDir = path.join(__dirname, "project");
-const inputProject = path.join(projectDir, "tsconfig.json");
+import { transpileProjectResult } from "./run";
 
 test("should transpile", () => {
-    const transpileResult = transpileProject(inputProject);
-
-    expect(transpileResult.diagnostics).not.toHaveDiagnostics();
-
-    // Check output paths relative to projectDir
-    const relativeResult = transpileResult.emitResult.map(({ name, text }) => ({
-        name: path.relative(projectDir, name),
-        text,
-    }));
-    expect(relativeResult).toMatchSnapshot();
+    const { diagnostics, emittedFiles } = transpileProjectResult(path.join(__dirname, "project", "tsconfig.json"));
+    expect(diagnostics).not.toHaveDiagnostics();
+    expect(emittedFiles).toMatchSnapshot();
 });