+++ test/unit/printer/sourcemaps.spec.ts
@@ -146,7 +146,7 @@ test.each([
 ])("Source map has correct mapping (%p)", async ({ code, assertPatterns }) => {
     const file = util.testModule(code).expectToHaveNoDiagnostics().getMainLuaFileResult();
 
-    const consumer = await new SourceMapConsumer(file.sourceMap);
+    const consumer = await new SourceMapConsumer(file.luaSourceMap);
     for (const { luaPattern, typeScriptPattern } of assertPatterns) {
         const luaPosition = lineAndColumnOf(file.lua, luaPattern);
         const mappedPosition = consumer.originalPositionFor(luaPosition);
@@ -190,11 +190,11 @@ test.each([
         .setMainFileName(fileName)
         .getMainLuaFileResult();
 
-    const sourceMap = JSON.parse(file.sourceMap);
+    const sourceMap = JSON.parse(file.luaSourceMap);
     expect(sourceMap.sources).toHaveLength(1);
     expect(sourceMap.sources[0]).toBe(mapSource);
 
-    const consumer = await new SourceMapConsumer(file.sourceMap);
+    const consumer = await new SourceMapConsumer(file.luaSourceMap);
     expect(consumer.sources).toHaveLength(1);
     expect(consumer.sources[0]).toBe(fullSource);
 });
@@ -212,7 +212,7 @@ test.each([
         .expectToHaveNoDiagnostics()
         .getMainLuaFileResult();
 
-    const sourceMap = JSON.parse(file.sourceMap);
+    const sourceMap = JSON.parse(file.luaSourceMap);
     expect(sourceMap.sourceRoot).toBe(mapSourceRoot);
 });
 
@@ -227,7 +227,7 @@ test.each([
 ])("Source map has correct name mappings (%p)", async ({ code, name }) => {
     const file = util.testModule(code).expectToHaveNoDiagnostics().getMainLuaFileResult();
 
-    const consumer = await new SourceMapConsumer(file.sourceMap);
+    const consumer = await new SourceMapConsumer(file.luaSourceMap);
     const typescriptPosition = lineAndColumnOf(code, name);
     let mappedName: string | undefined;
     consumer.eachMapping(mapping => {
@@ -289,14 +289,15 @@ test("Inline sourcemaps", () => {
 
     const file = util
         .testFunction(code)
-        .setOptions({ inlineSourceMap: true })
+        .setOptions({ inlineSourceMap: true }) // We can't disable 'sourceMap' option because it's used for comparison
+        .disableSemanticCheck() // TS5053: Option 'sourceMap' cannot be specified with option 'inlineSourceMap'.
         .expectToMatchJsResult()
         .getMainLuaFileResult();
 
     const [, inlineSourceMapMatch] =
         file.lua.match(/--# sourceMappingURL=data:application\/json;base64,([A-Za-z0-9+/=]+)/) ?? [];
     const inlineSourceMap = Buffer.from(inlineSourceMapMatch, "base64").toString();
-    expect(file.sourceMap).toBe(inlineSourceMap);
+    expect(inlineSourceMap).toBe(file.luaSourceMap);
 });
 
 // Helper functions